Migrate ava cli dev tests to jest
