--- .github/workflows/test-integration-dev.yml
@@ -1,47 +1,0 @@
-name: Dev
-
-on:
-  push:
-    branches:
-    - main
-    tags:
-    - '!*'
-  pull_request:
-
-jobs:
-  test:
-    name: Dev
-    timeout-minutes: 75
-    strategy:
-      fail-fast: false
-      matrix:
-        os: [ubuntu-latest, macos-latest]
-        node: [12]
-    runs-on: ${{ matrix.os }}
-    env:
-      TURBO_REMOTE_ONLY: true
-      TURBO_TEAM: vercel
-      TURBO_TOKEN: ${{ secrets.TURBO_TOKEN }}
-    steps:
-      - uses: actions/setup-go@v2
-        with:
-          go-version: '1.13.15'
-      - uses: actions/setup-node@v2
-        with:
-          node-version: ${{ matrix.node }}
-      - uses: actions/checkout@v2
-        with:
-          fetch-depth: 100
-      - run: git --version
-      - run: git fetch origin main --depth=100
-      - run: git fetch origin ${{ github.ref }} --depth=100
-      - run: git diff origin/main...HEAD --name-only
-      - name: Install Hugo
-        if: matrix.os == 'macos-latest'
-        run: curl -L -O https://github.com/gohugoio/hugo/releases/download/v0.56.0/hugo_0.56.0_macOS-64bit.tar.gz && tar -xzf hugo_0.56.0_macOS-64bit.tar.gz && mv ./hugo packages/cli/test/dev/fixtures/08-hugo/
-      - run: yarn install --network-timeout 1000000
-      - run: yarn run build
-      - run: yarn test-integration-dev
-        env:
-          VERCEL_TEAM_TOKEN: ${{ secrets.VERCEL_TEAM_TOKEN }}
-          VERCEL_REGISTRATION_URL: ${{ secrets.VERCEL_REGISTRATION_URL }}
