+++ packages/cli/package.json
@@ -13,9 +13,9 @@
   "scripts": {
     "preinstall": "node ./scripts/preinstall.js",
     "test": "jest --env node --verbose --runInBand --bail",
-    "test-unit": "yarn test",
+    "test-unit": "yarn test test/unit/",
     "test-integration-cli": "rimraf test/fixtures/integration && ava test/integration.js --serial --fail-fast --verbose",
-    "test-integration-dev": "ava test/dev/integration.js --serial --fail-fast --verbose",
+    "test-integration-dev": "yarn test test/dev/",
     "prepublishOnly": "yarn build",
     "coverage": "codecov",
     "build": "node -r ts-eager/register ./scripts/build.ts",
