+++ packages/cli/test/unit/commands/whoami.test.ts
@@ -1,6 +1,6 @@
-import { client } from '../mocks/client';
-import { useUser } from '../mocks/user';
-import whoami from '../../src/commands/whoami';
+import { client } from '../../mocks/client';
+import { useUser } from '../../mocks/user';
+import whoami from '../../../src/commands/whoami';
 
 describe('whoami', () => {
   it('should reject invalid arguments', async () => {
