+++ packages/cli/test/unit/util/get-project-name.test.ts
@@ -1,4 +1,4 @@
-import getProjectName from '../../src/util/get-project-name';
+import getProjectName from '../../../src/util/get-project-name';
 
 describe('getProjectName', () => {
   it('should work with argv', () => {
