+++ packages/cli/test/unit/util/get-update-command.test.ts
@@ -1,5 +1,5 @@
-import { isCanary } from '../../src/util/is-canary';
-import getUpdateCommand from '../../src/util/get-update-command';
+import { isCanary } from '../../../src/util/is-canary';
+import getUpdateCommand from '../../../src/util/get-update-command';
 
 describe('getUpdateCommand', () => {
   it('should detect update command', async () => {
