+++ packages/cli/test/unit/util/init/did-you-mean.test.ts
@@ -1,4 +1,4 @@
-import didYouMean from '../../../src/util/init/did-you-mean';
+import didYouMean from '../../../../src/util/init/did-you-mean';
 
 describe('didYouMean', () => {
   const examples = [
