+++ packages/cli/test/unit/util/is-valid-name.test.ts
@@ -1,4 +1,4 @@
-import { isValidName } from '../../src/util/is-valid-name';
+import { isValidName } from '../../../src/util/is-valid-name';
 
 const tests = {
   'hello world': true,
