+++ packages/cli/test/unit/util/projects/get-vercel-directory.test.ts
@@ -1,5 +1,5 @@
 import { basename, join } from 'path';
-import { getVercelDirectory } from '../../../src/util/projects/link';
+import { getVercelDirectory } from '../../../../src/util/projects/link';
 
 const fixture = (name: string) => join(__dirname, '../../fixtures/unit', name);
